+++ app/src/main/java/com/hhp227/yu_minigroup/activity/ModifyActivity.java
@@ -1,10 +1,9 @@
-package com.hhp227.yu_minigroup;
+package com.hhp227.yu_minigroup.activity;
 
 import android.app.ProgressDialog;
 import android.content.Intent;
 import android.graphics.Bitmap;
 import android.net.Uri;
-import android.os.Bundle;
 import android.os.Environment;
 import android.provider.MediaStore;
 import android.text.Html;
@@ -13,28 +12,24 @@ import android.text.TextUtils;
 import android.util.Log;
 import android.view.*;
 import android.widget.Toast;
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
+import android.os.Bundle;
 import androidx.core.content.FileProvider;
 import androidx.exifinterface.media.ExifInterface;
 import androidx.recyclerview.widget.LinearLayoutManager;
-import com.android.volley.Request;
-import com.android.volley.VolleyLog;
+import com.android.volley.*;
 import com.android.volley.toolbox.StringRequest;
-import com.android.volley.toolbox.Volley;
 import com.google.android.material.snackbar.Snackbar;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.*;
+import com.hhp227.yu_minigroup.R;
 import com.hhp227.yu_minigroup.adapter.WriteListAdapter;
 import com.hhp227.yu_minigroup.app.AppController;
 import com.hhp227.yu_minigroup.app.EndPoint;
-import com.hhp227.yu_minigroup.databinding.ActivityWriteBinding;
+import com.hhp227.yu_minigroup.dto.ArticleItem;
 import com.hhp227.yu_minigroup.dto.YouTubeItem;
 import com.hhp227.yu_minigroup.helper.BitmapUtil;
-import com.hhp227.yu_minigroup.helper.PreferenceManager;
 import com.hhp227.yu_minigroup.volley.util.MultipartRequest;
-import net.htmlparser.jericho.Source;
-import org.json.JSONException;
-import org.json.JSONObject;
 
 import java.io.ByteArrayOutputStream;
 import java.io.File;
@@ -42,24 +37,16 @@ import java.io.IOException;
 import java.text.SimpleDateFormat;
 import java.util.*;
 
-public class WriteActivity extends AppCompatActivity {
-    public static final int CAMERA_PICK_IMAGE_REQUEST_CODE = 100;
+import static com.hhp227.yu_minigroup.activity.WriteActivity.*;
 
-    public static final int REQUEST_IMAGE_CAPTURE = 200;
+public class ModifyActivity extends AppCompatActivity {
+    private static final String TAG = ModifyActivity.class.getSimpleName();
 
-    public static final int REQUEST_YOUTUBE_PICK = 300;
-
-    private static final String TAG = WriteActivity.class.getSimpleName();
-
-    private boolean mIsAdmin;
-
-    private String mGrpId, mGrpNm, mGrpImg, mCurrentPhotoPath, mCookie, mKey;
+    private String mGrpId, mArtlNum, mCurrentPhotoPath, mCookie, mTitle, mContent, mGrpKey, mArtlKey;
 
     private List<String> mImageList;
 
     private List<Object> mContents;
-
-    private PreferenceManager mPreferenceManager;
 
     private ProgressDialog mProgressDialog;
 
@@ -79,27 +66,39 @@ public class WriteActivity extends AppCompatActivity {
         mBinding = ActivityWriteBinding.inflate(getLayoutInflater());
 
         setContentView(mBinding.getRoot());
+        Intent intent = getIntent();
+        Map<String, Object> headerMap = new HashMap<>();
         mContents = new ArrayList<>();
         mAdapter = new WriteListAdapter(mContents);
-        mPreferenceManager = AppController.getInstance().getPreferenceManager();
         mCookie = AppController.getInstance().getCookieManager().getCookie(EndPoint.LOGIN_LMS);
         mProgressDialog = new ProgressDialog(this);
-        mIsAdmin = getIntent().getBooleanExtra("admin", false);
-        mGrpId = getIntent().getStringExtra("grp_id");
-        mGrpNm = getIntent().getStringExtra("grp_nm");
-        mGrpImg = getIntent().getStringExtra("grp_img");
-        mKey = getIntent().getStringExtra("key");
+        mGrpId = intent.getStringExtra("grp_id");
+        mArtlNum = intent.getStringExtra("artl_num");
+        mTitle = intent.getStringExtra("sbjt");
+        mContent = intent.getStringExtra("txt");
+        mImageList = intent.getStringArrayListExtra("img");
+        mYouTubeItem = intent.getParcelableExtra("vid");
+        mGrpKey = intent.getStringExtra("grp_key");
+        mArtlKey = intent.getStringExtra("artl_key");
 
         setSupportActionBar(mBinding.toolbar);
         if (getSupportActionBar() != null) {
             getSupportActionBar().setDisplayHomeAsUpEnabled(true);
         }
-        mAdapter.addHeaderView(new HashMap<>());
-        mBinding.rvWrite.setLayoutManager(new LinearLayoutManager(this));
-        mBinding.rvWrite.setAdapter(mAdapter);
         mBinding.llImage.setOnClickListener(this::showContextMenu);
         mBinding.llVideo.setOnClickListener(this::showContextMenu);
+        headerMap.put("title", mTitle);
+        headerMap.put("content", mContent);
+        mAdapter.addHeaderView(headerMap);
+        mBinding.rvWrite.setLayoutManager(new LinearLayoutManager(this));
+        mBinding.rvWrite.setAdapter(mAdapter);
         mProgressDialog.setCancelable(false);
+        if (mImageList.size() > 0) {
+            mContents.addAll(mImageList);
+            mAdapter.notifyDataSetChanged();
+        }
+        if (mYouTubeItem != null)
+            mContents.add(mYouTubeItem.position + 1, mYouTubeItem);
     }
 
     @Override
@@ -126,25 +125,29 @@ public class WriteActivity extends AppCompatActivity {
 
                 if (!title.isEmpty() && !(TextUtils.isEmpty(content) && mContents.size() < 2)) {
                     mMakeHtmlContents = new StringBuilder();
-                    mImageList = new ArrayList<>();
 
+                    mImageList.clear();
                     mProgressDialog.setMessage("전송중...");
                     mProgressDialog.setProgressStyle(mContents.size() > 1 ? ProgressDialog.STYLE_HORIZONTAL : ProgressDialog.STYLE_SPINNER);
                     showProgressDialog();
                     if (mContents.size() > 1) {
                         int position = 1;
 
-                        if (mContents.get(position) instanceof Bitmap) {////////////// 리팩토링 요망
-                            Bitmap bitmap = (Bitmap) mContents.get(position);// 수정
+                        if (mContents.get(position) instanceof String) {
+                            String image = (String) mContents.get(position);
 
-                            uploadImage(position, bitmap); // 수정
+                            uploadProcess(position, image, false);
+                        } else if (mContents.get(position) instanceof Bitmap) {
+                            Bitmap bitmap = (Bitmap) mContents.get(position);
+
+                            uploadImage(position, bitmap);//
                         } else if (mContents.get(position) instanceof YouTubeItem) {
                             YouTubeItem youTubeItem = (YouTubeItem) mContents.get(position);
 
                             uploadProcess(position, youTubeItem.videoId, true);
                         }
                     } else
-                        actionSend(mGrpId, title, content);
+                        actionSend(title, content);
                 } else
                     Snackbar.make(getCurrentFocus(), (title.isEmpty() ? "제목" : "내용") + "을 입력하세요.", Snackbar.LENGTH_LONG).setAction("Action", null).show();
                 return true;
@@ -174,10 +177,12 @@ public class WriteActivity extends AppCompatActivity {
 
         switch (item.getGroupId()) {
             case 0:
-                if (mContents.get(item.getItemId()) instanceof YouTubeItem)
+                int position = item.getItemId();
+
+                if (mContents.get(position) instanceof YouTubeItem)
                     mYouTubeItem = null;
-                mContents.remove(item.getItemId());
-                mAdapter.notifyItemRemoved(item.getItemId());
+                mContents.remove(position);
+                mAdapter.notifyItemRemoved(position);
                 return true;
             case 1:
                 intent = new Intent(Intent.ACTION_PICK);
@@ -211,7 +216,7 @@ public class WriteActivity extends AppCompatActivity {
                 else {
                     Intent ysIntent = new Intent(getApplicationContext(), YouTubeSearchActivity.class);
 
-                    ysIntent.putExtra("type", 0);
+                    ysIntent.putExtra("type", 1);
                     startActivityForResult(ysIntent, REQUEST_YOUTUBE_PICK);
                 }
                 return true;
@@ -257,14 +262,14 @@ public class WriteActivity extends AppCompatActivity {
         }
     }
 
-    private void uploadImage(int position, Bitmap bitmap) {// 수정
+    private void uploadImage(int position, Bitmap bitmap) { // 수정
         MultipartRequest multipartRequest = new MultipartRequest(Request.Method.POST, EndPoint.IMAGE_UPLOAD, response -> {
             String imageSrc = new String(response.data);
             imageSrc = EndPoint.BASE_URL + imageSrc.substring(imageSrc.lastIndexOf("/ilosfiles/"), imageSrc.lastIndexOf("\""));
 
             uploadProcess(position, imageSrc, false);
         }, error -> {
-            VolleyLog.e(error.getMessage());
+            VolleyLog.e(TAG, error.getMessage());
             hideProgressDialog();
         }) {
             @Override
@@ -279,7 +284,7 @@ public class WriteActivity extends AppCompatActivity {
             protected Map<String, DataPart> getByteData() {
                 Map<String, DataPart> params = new HashMap<>();
 
-                params.put("file", new DataPart(System.currentTimeMillis() + position + ".jpg", getFileDataFromDrawable(bitmap)));
+                params.put("file", new MultipartRequest.DataPart(System.currentTimeMillis() + position + ".jpg", getFileDataFromDrawable(bitmap))); // 수정
                 return params;
             }
 
@@ -290,27 +295,31 @@ public class WriteActivity extends AppCompatActivity {
                 return byteArrayOutputStream.toByteArray();
             }
         };
-        Volley.newRequestQueue(this).add(multipartRequest);
+        AppController.getInstance().addToRequestQueue(multipartRequest);
     }
 
-    private void uploadProcess(int position, String imageSrc, boolean isYoutube) { // 추가
+    private void uploadProcess(int position, String imageUrl, boolean isYoutube) { // 수정
         if (!isYoutube)
-            mImageList.add(imageSrc);
+            mImageList.add(imageUrl);
         mProgressDialog.setProgress((int) ((double) (position) / (mContents.size() - 1) * 100));
         try {
-            String test = (isYoutube ? "<p><embed title=\"YouTube video player\" class=\"youtube-player\" autostart=\"true\" src=\"//www.youtube.com/embed/" + imageSrc + "?autoplay=1\"  width=\"488\" height=\"274\"></embed><p>" // 유튜브 태그
-                    : ("<p><img src=\"" + imageSrc + "\" width=\"488\"><p>")) + (position < mContents.size() - 1 ? "<br>": "");
+            String test = (isYoutube ? "<p><embed title=\"YouTube video player\" class=\"youtube-player\" autostart=\"true\" src=\"//www.youtube.com/embed/" + imageUrl + "?autoplay=1\"  width=\"488\" height=\"274\"></embed><p>" // 유튜브 태그
+                    : ("<p><img src=\"" + imageUrl + "\" width=\"488\"><p>")) + (position < mContents.size() - 1 ? "<br>": "");
 
             mMakeHtmlContents.append(test);
             if (position < mContents.size() - 1) {
                 position++;
-                Thread.sleep(isYoutube ? 0 : 700);
+                Thread.sleep(!isYoutube ? 700 : 0);
 
                 // 분기
-                if (mContents.get(position) instanceof Bitmap) {
-                    Bitmap bitmap = (Bitmap) mContents.get(position);
+                if (mContents.get(position) instanceof Bitmap) {////////////// 리팩토링 요망
+                    Bitmap bitmap = (Bitmap) mContents.get(position);// 수정
 
-                    uploadImage(position, bitmap);
+                    uploadImage(position, bitmap); // 수정
+                } else if (mContents.get(position) instanceof String) {
+                    String imageSrc = (String) mContents.get(position);
+
+                    uploadProcess(position, imageSrc, false);
                 } else if (mContents.get(position) instanceof YouTubeItem) {
                     YouTubeItem youTubeItem = (YouTubeItem) mContents.get(position);
 
@@ -320,41 +329,29 @@ public class WriteActivity extends AppCompatActivity {
                 String title = (String) mAdapter.getTextMap().get("title");
                 String content = (!TextUtils.isEmpty(mAdapter.getTextMap().get("content").toString()) ? Html.toHtml((Spanned) mAdapter.getTextMap().get("content"), Html.TO_HTML_PARAGRAPH_LINES_INDIVIDUAL) + "<p><br data-mce-bogus=\"1\"></p>" : "") + mMakeHtmlContents.toString();
 
-                actionSend(mGrpId, title, content);
+                actionSend(title, content);
             }
-        } catch (Exception e) {
+        } catch (InterruptedException e) {
             e.printStackTrace();
-            Snackbar.make(getCurrentFocus(), "이미지 업로드 실패", Snackbar.LENGTH_LONG).setAction("Action", null).show();
+            Toast.makeText(getApplicationContext(), "이미지 업로드 실패", Toast.LENGTH_LONG).show();
             hideProgressDialog();
         }
     }
 
-    private void actionSend(String grpId, String title, String content) {
+    private void actionSend(String title, String content) {
         String tagStringReq = "req_send";
-        StringRequest stringRequest = new StringRequest(Request.Method.POST, EndPoint.WRITE_ARTICLE, response -> {
+        StringRequest stringRequest = new StringRequest(Request.Method.POST, EndPoint.MODIFY_ARTICLE, response -> {
             hideProgressDialog();
             try {
-                JSONObject jsonObject = new JSONObject(response);
-                boolean error = jsonObject.getBoolean("isError");
+                Intent intent = new Intent(ModifyActivity.this, ArticleActivity.class);
 
-                if (!error) {
-                    Intent intent = new Intent(WriteActivity.this, GroupActivity.class);
-
-                    intent.putExtra("admin", mIsAdmin);
-                    intent.putExtra("grp_id", grpId);
-                    intent.putExtra("grp_nm", mGrpNm);
-                    intent.putExtra("grp_img", mGrpImg);
-                    intent.putExtra("key", mKey);
-
-                    // 이전 Activity 초기화
-                    intent.setFlags(Intent.FLAG_ACTIVITY_TASK_ON_HOME | Intent.FLAG_ACTIVITY_CLEAR_TOP);
-                    startActivity(intent);
-                    Toast.makeText(getApplicationContext(), "전송완료", Toast.LENGTH_LONG).show();
-                }
-            } catch (JSONException e) {
-                Log.e(TAG, "에러 : " + e.getMessage());
+                Toast.makeText(getApplicationContext(), "수정완료", Toast.LENGTH_LONG).show();
+                setResult(RESULT_OK, intent);
+                finish();
+            } catch (Exception e) {
+                Log.e(TAG, e.getMessage());
             } finally {
-                getArticleId();
+                initFirebaseData();
             }
         }, error -> {
             VolleyLog.e(error.getMessage());
@@ -373,32 +370,14 @@ public class WriteActivity extends AppCompatActivity {
             protected Map<String, String> getParams() {
                 Map<String, String> params = new HashMap<>();
 
+                params.put("CLUB_GRP_ID", mGrpId);
+                params.put("ARTL_NUM", mArtlNum);
                 params.put("SBJT", title);
-                params.put("CLUB_GRP_ID", grpId);
                 params.put("TXT", content);
                 return params;
             }
         };
         AppController.getInstance().addToRequestQueue(stringRequest, tagStringReq);
-    }
-
-    private void getArticleId() {
-        String params = "?CLUB_GRP_ID=" + mGrpId + "&displayL=1";
-
-        AppController.getInstance().addToRequestQueue(new StringRequest(Request.Method.GET, EndPoint.GROUP_ARTICLE_LIST + params, response -> {
-            Source source = new Source(response);
-            String artlNum = source.getFirstElementByClass("comment_wrap").getAttributeValue("num");
-
-            insertArticleToFirebase(artlNum);
-        }, error -> VolleyLog.e(error.getMessage())) {
-            @Override
-            public Map<String, String> getHeaders() {
-                Map<String, String> headers = new HashMap<>();
-
-                headers.put("Cookie", mCookie);
-                return headers;
-            }
-        });
     }
 
     private File createImageFile() throws IOException {
@@ -411,6 +390,7 @@ public class WriteActivity extends AppCompatActivity {
                 storageDir      /* directory */
         );
         mCurrentPhotoPath = image.getAbsolutePath();
+
         return image;
     }
 
@@ -420,19 +400,32 @@ public class WriteActivity extends AppCompatActivity {
         unregisterForContextMenu(v);
     }
 
-    private void insertArticleToFirebase(String artlNum) {
+    private void initFirebaseData() {
         DatabaseReference databaseReference = FirebaseDatabase.getInstance().getReference("Articles");
-        Map<String, Object> map = new HashMap<>();
 
-        map.put("id", artlNum);
-        map.put("uid", mPreferenceManager.getUser().getUid());
-        map.put("name", mPreferenceManager.getUser().getName());
-        map.put("title", mAdapter.getTextMap().get("title"));
-        map.put("timestamp", System.currentTimeMillis());
-        map.put("content", TextUtils.isEmpty(mAdapter.getTextMap().get("content").toString()) ? null : mAdapter.getTextMap().get("content").toString());
-        map.put("images", mImageList);
-        map.put("youtube", mYouTubeItem);
-        databaseReference.child(mKey).push().setValue(map);
+        updateArticleDataToFirebase(databaseReference.child(mGrpKey).child(mArtlKey));
+    }
+
+    private void updateArticleDataToFirebase(DatabaseReference query) {
+        query.addListenerForSingleValueEvent(new ValueEventListener() {
+            @Override
+            public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
+                ArticleItem articleItem = dataSnapshot.getValue(ArticleItem.class);
+
+                if (articleItem != null) {
+                    articleItem.setTitle((String) mAdapter.getTextMap().get("title"));
+                    articleItem.setContent(TextUtils.isEmpty(mAdapter.getTextMap().get("content").toString()) ? null : mAdapter.getTextMap().get("content").toString());
+                    articleItem.setImages(mImageList.isEmpty() ? null : mImageList);
+                    articleItem.setYoutube(mYouTubeItem);
+                    query.getRef().setValue(articleItem);
+                }
+            }
+
+            @Override
+            public void onCancelled(@NonNull DatabaseError databaseError) {
+                Log.e("파이어베이스", databaseError.getMessage());
+            }
+        });
     }
 
     private void showProgressDialog() {
