+++ app/src/main/java/com/hhp227/yu_minigroup/fragment/GroupFragment.java
@@ -11,12 +11,10 @@ import android.util.Log;
 import android.util.TypedValue;
 import android.view.WindowManager;
 import android.webkit.CookieManager;
-import android.webkit.ValueCallback;
-import android.widget.ProgressBar;
+
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.ActionBarDrawerToggle;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.Toolbar;
 import androidx.drawerlayout.widget.DrawerLayout;
 import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
@@ -25,18 +23,21 @@ import android.view.ViewGroup;
 
 import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
-import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
+
 import com.android.volley.Request;
 import com.android.volley.VolleyLog;
 import com.android.volley.toolbox.StringRequest;
-import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.google.firebase.database.*;
-import com.hhp227.yu_minigroup.*;
 import com.hhp227.yu_minigroup.R;
+import com.hhp227.yu_minigroup.activity.CreateActivity;
+import com.hhp227.yu_minigroup.activity.FindActivity;
+import com.hhp227.yu_minigroup.activity.GroupActivity;
+import com.hhp227.yu_minigroup.activity.LoginActivity;
+import com.hhp227.yu_minigroup.activity.MainActivity;
+import com.hhp227.yu_minigroup.activity.RequestActivity;
 import com.hhp227.yu_minigroup.adapter.GroupGridAdapter;
 import com.hhp227.yu_minigroup.app.AppController;
 import com.hhp227.yu_minigroup.app.EndPoint;
-import com.hhp227.yu_minigroup.databinding.FragmentGroupBinding;
 import com.hhp227.yu_minigroup.dto.GroupItem;
 import com.hhp227.yu_minigroup.helper.PreferenceManager;
 import net.htmlparser.jericho.Element;
