+++ app/src/main/java/com/hhp227/yu_minigroup/fragment/SeatFragment.java
@@ -15,12 +15,11 @@ import androidx.recyclerview.widget.LinearLayoutManager;
 import com.android.volley.Request;
 import com.android.volley.VolleyLog;
 import com.android.volley.toolbox.JsonObjectRequest;
-import com.hhp227.yu_minigroup.MainActivity;
+import com.hhp227.yu_minigroup.activity.MainActivity;
 import com.hhp227.yu_minigroup.R;
 import com.hhp227.yu_minigroup.adapter.SeatListAdapter;
 import com.hhp227.yu_minigroup.app.AppController;
 import com.hhp227.yu_minigroup.app.EndPoint;
-import com.hhp227.yu_minigroup.databinding.FragmentSeatBinding;
 import com.hhp227.yu_minigroup.dto.SeatItem;
 import org.json.JSONArray;
 import org.json.JSONException;
