+++ app/src/main/java/com/hhp227/yu_minigroup/fragment/Tab1Fragment.java
@@ -19,13 +19,12 @@ import com.android.volley.Request;
 import com.android.volley.VolleyLog;
 import com.android.volley.toolbox.StringRequest;
 import com.google.firebase.database.*;
-import com.hhp227.yu_minigroup.ArticleActivity;
+import com.hhp227.yu_minigroup.activity.ArticleActivity;
 import com.hhp227.yu_minigroup.R;
-import com.hhp227.yu_minigroup.WriteActivity;
+import com.hhp227.yu_minigroup.activity.WriteActivity;
 import com.hhp227.yu_minigroup.adapter.ArticleListAdapter;
 import com.hhp227.yu_minigroup.app.AppController;
 import com.hhp227.yu_minigroup.app.EndPoint;
-import com.hhp227.yu_minigroup.databinding.FragmentTab1Binding;
 import com.hhp227.yu_minigroup.dto.ArticleItem;
 import com.hhp227.yu_minigroup.dto.YouTubeItem;
 import net.htmlparser.jericho.Element;
