+++ app/src/main/java/com/hhp227/yu_minigroup/fragment/Tab4Fragment.java
@@ -28,11 +28,14 @@ import com.bumptech.glide.load.model.LazyHeaders;
 import com.bumptech.glide.request.RequestOptions;
 import com.google.android.gms.ads.AdRequest;
 import com.google.firebase.database.*;
-import com.hhp227.yu_minigroup.*;
 import com.hhp227.yu_minigroup.R;
+import com.hhp227.yu_minigroup.activity.MainActivity;
+import com.hhp227.yu_minigroup.activity.NoticeActivity;
+import com.hhp227.yu_minigroup.activity.ProfileActivity;
+import com.hhp227.yu_minigroup.activity.SettingsActivity;
+import com.hhp227.yu_minigroup.activity.VerInfoActivity;
 import com.hhp227.yu_minigroup.app.AppController;
 import com.hhp227.yu_minigroup.app.EndPoint;
-import com.hhp227.yu_minigroup.databinding.ContentTab4Binding;
 import com.hhp227.yu_minigroup.databinding.FragmentTab4Binding;
 import com.hhp227.yu_minigroup.dto.GroupItem;
 import com.hhp227.yu_minigroup.dto.User;
