+++ app/src/main/java/com/hhp227/yu_minigroup/fragment/TabHostLayoutFragment.java
@@ -18,8 +18,7 @@ import com.bumptech.glide.Glide;
 import com.bumptech.glide.request.RequestOptions;
 import com.google.android.material.tabs.TabLayout;
 import com.hhp227.yu_minigroup.R;
-import com.hhp227.yu_minigroup.WriteActivity;
-import com.hhp227.yu_minigroup.databinding.FragmentTabHostLayoutBinding;
+import com.hhp227.yu_minigroup.activity.WriteActivity;
 
 import java.util.Arrays;
 import java.util.List;
