+++ app/src/main/res/layout/activity_article.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ArticleActivity">
+        tools:context=".activity.ArticleActivity">
 
     <com.google.android.material.appbar.AppBarLayout
             android:layout_height="wrap_content"
