+++ app/src/main/res/layout/activity_group.xml
@@ -5,6 +5,6 @@
         android:id="@+id/content_frame"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".GroupActivity">
+        tools:context=".activity.GroupActivity">
 
 </FrameLayout>
