+++ app/src/main/res/layout/activity_picture.xml
@@ -6,7 +6,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@android:color/black"
-        tools:context=".PictureActivity">
+        tools:context=".activity.PictureActivity">
 
     <com.google.android.material.appbar.AppBarLayout
             android:layout_height="wrap_content"
