+++ app/src/main/res/layout/activity_reply_modify.xml
@@ -5,7 +5,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ReplyModifyActivity">
+        tools:context=".activity.ReplyModifyActivity">
 
     <com.google.android.material.appbar.AppBarLayout
             android:layout_height="wrap_content"
