+++ app/src/main/res/layout/activity_splash.xml
@@ -6,7 +6,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/colorPrimary"
-        tools:context=".SplashActivity">
+        tools:context=".activity.SplashActivity">
 
     <ImageView
         android:id="@+id/ImageView1"
