+++ app/src/main/res/layout/activity_ver_info.xml
@@ -5,7 +5,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".VerInfoActivity">
+        tools:context=".activity.VerInfoActivity">
 
     <com.google.android.material.appbar.AppBarLayout
             android:layout_height="wrap_content"
