+++ app/src/main/res/layout/activity_web_view.xml
@@ -5,4 +5,4 @@
         android:id="@+id/wv_news"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".WebViewActivity"/>
+        tools:context=".activity.WebViewActivity"/>
