+++ app/src/main/res/layout/schedule_item.xml
@@ -4,7 +4,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        tools:context=".MainActivity">
+        tools:context=".activity.MainActivity">
 
     <LinearLayout
             android:layout_width="match_parent"
