+++ app/src/main/res/menu/main.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
       xmlns:app="http://schemas.android.com/apk/res-auto" xmlns:tools="http://schemas.android.com/tools"
-      tools:context=".GroupActivity">
+      tools:context=".activity.GroupActivity">
 
     <!-- Chat -->
     <item android:id="@+id/action_chat"
