+++ app/src/main/res/menu/menu_create.xml
@@ -3,7 +3,7 @@
         xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:tools="http://schemas.android.com/tools"
         xmlns:app="http://schemas.android.com/apk/res-auto"
-        tools:context=".WriteActivity">
+        tools:context=".activity.WriteActivity">
 
     <item android:id="@+id/action_send"
           android:title="생성"
