+++ app/src/main/res/menu/write.xml
@@ -3,7 +3,7 @@
         xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
         xmlns:tools="http://schemas.android.com/tools"
-        tools:context=".WriteActivity">
+        tools:context=".activity.WriteActivity">
 
     <item android:id="@+id/action_send"
           android:title="전송"
